+++ src/main/kotlin/Main.kt
@@ -1,9 +1,10 @@
+import presenter.PresenterImpl
 import view.ViewImpl
 
 /**
  * Created by yurich on 02.12.16.
  */
-fun main(args: Array<String>) {
-    val view = ViewImpl()
+fun main() {
+    val view = ViewImpl.create { view -> PresenterImpl(view) }
     view.play()
 }